:sparkles:	feat Feature/seat update

+++ src/main/java/com/example/momentix/domain/events/dto/response/SeatResponseDto.java
@@ -14,8 +14,8 @@ public class SeatResponseDto {
     @CsvBindByName
     private String seatPartType;
     @CsvBindByName
-    private int seatRow;
+    private Long seatRow;
     @CsvBindByName
-    private int seatCol;
+    private Long seatCol;
 
 }
